igalapedia/requirements.txt
asgiref==3.7.2
Django==5.0.3
django-filter==24.2
psycopg2-binary==2.9.10
sqlparse==0.4.4
tzdata==2024.1
whitenoise==6.6.0
python-dotenv==1.0.1
dj-database-url==1.0.0
gunicorn==23.0.0
django-storages==1.14.4
boto3==1.35.0
Pillow==11.0.0
django-ckeditor==6.7.0
